Moving views to a separate directory
