+++ db_create.py
@@ -3,7 +3,7 @@ import os.path
 
 from migrate.versioning import api
 
-from app import db
+from pears import db
 from config import SQLALCHEMY_DATABASE_URI
 from config import SQLALCHEMY_MIGRATE_REPO
