+++ db_migrate.py
@@ -4,7 +4,7 @@ import imp
 
 from migrate.versioning import api
 
-from app import db
+from pears import db
 from config import SQLALCHEMY_DATABASE_URI
 from config import SQLALCHEMY_MIGRATE_REPO
