+++ db_repository/versions/001_migration.py
@@ -1,0 +1,27 @@
+from sqlalchemy import *
+from migrate import *
+
+
+from migrate.changeset import schema
+pre_meta = MetaData()
+post_meta = MetaData()
+open_vectors = Table('open_vectors', post_meta,
+    Column('id', Integer, primary_key=True, nullable=False),
+    Column('word', UnicodeText(length=64)),
+    Column('vector', Text),
+)
+
+
+def upgrade(migrate_engine):
+    # Upgrade operations go here. Don't create your own engine; bind
+    # migrate_engine to your metadata
+    pre_meta.bind = migrate_engine
+    post_meta.bind = migrate_engine
+    post_meta.tables['open_vectors'].create()
+
+
+def downgrade(migrate_engine):
+    # Operations to reverse the above upgrade go here.
+    pre_meta.bind = migrate_engine
+    post_meta.bind = migrate_engine
+    post_meta.tables['open_vectors'].drop()
