+++ pears/examples/demo1pear/doc.dists
