+++ pears/examples/demo2pear/doc.dists
