+++ pears/runTextBlobTagger.py
