+++ pears/static/css/simple-sidebar.css
