+++ pears/static/favicon.png
