+++ pears/static/js/jquery.js
