+++ pears/static/pears-logo-small.png
