+++ pears/templates/index.html
@@ -21,9 +21,9 @@
     <div class="col-lg-12 text-center v-center">
 
       <br><br><br>
-      <a href="{{ url_for('index')}}"><img src="{{ url_for('static', filename='pears-logo-small.png')}}"></a>
+      <a href="{{ url_for('.index')}}"><img src="{{ url_for('static', filename='pears-logo-small.png')}}"></a>
       <br/><br/>
-      <form class="col-lg-12" action="{{url_for('index')}}">
+      <form class="col-lg-12" action="{{url_for('.index')}}">
         <div class="input-group" style="width:340px;text-align:center;margin:0 auto;">
           <input class="form-control input-lg" title="Enter your Search Query." placeholder="Enter your Search Query"
                  type="text" name='q'>
