+++ pears/templates/results.html
@@ -25,7 +25,7 @@
         <span class="icon-bar"></span>
         <span class="icon-bar"></span>
       </button>
-      <a class="navbar-brand" href="{{ url_for('index')}}"><img
+      <a class="navbar-brand" href="{{ url_for('.index')}}"><img
         src="{{ url_for('static', filename='pears-logo-small.png')}}" height="40px"></a>
       </a>
     </div>
