+++ pears/views/main.py
@@ -1,16 +1,16 @@
 #! /usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from flask import render_template, request
+from flask import render_template, request, Blueprint
 
-import best_pears
-import scorePages
-from app import app
-from app.utils import read_pears, query_distribution, load_entropies
+from pears import best_pears
+from pears import scorePages
+from pears.utils import read_pears, query_distribution, load_entropies
 
+mod = Blueprint('main', __name__)
 
-@app.route('/')
-@app.route('/index')
+@mod.route('/')
+@mod.route('/index')
 def index():
     pages = []
     entropies_dict = load_entropies()
@@ -30,7 +30,6 @@ def index():
             pear_names = []
             for p in pears:
                 pear_names.append(p[0])
-                print p
             pages = scorePages.runScript(query, query_dist, pear_names)
 
         # '''remove the following lines after testing'''
