+++ run.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-from app import app
+from pears import app
 
 if __name__ == "__main__":
     app.run(debug=True, host='0.0.0.0')
